Cleaned up the Output.ts file a bit. Still a long way to go.
