+++ src/Output/external/diagram.js
